triposr-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==1.10.13
numpy==1.26.4
Pillow==10.1.0
einops==0.7.0
transformers==4.35.0
trimesh==4.0.5
rembg==2.0.53
huggingface-hub==0.22.2
imageio[ffmpeg]==2.34.0
xatlas==0.0.9
moderngl==5.10.0
torch==2.2.2
torchvision==0.17.2
git+https://github.com/VAST-AI-Research/TripoSR.git
git+https://github.com/tatsy/torchmcubes.git
